requirements.txt
Flask
flask_cors
pyyaml
transformers